Move closestAttributeValue into its own module
Allows the polyfill it requires to be tree-shaken properly
when components that don't use it import `common.mjs`
